1. Separated styles and components from the login.tsx file.
