--- src/Login/LoginStyle.css
@@ -1,3 +1,0 @@
-.login-form{
-
-}
